removed orig Signature, replaced w/class-based version
