+++ routes/articles.js
@@ -2,7 +2,6 @@ const express = require('express');
 const router = express.Router();
 const Article = require('../modules/Article');
 const Signature = require('../modules/Signature');
-const Signature2 = require('../modules/SignatureClass');
 const debug = require('debug')('routes:articles');
 
 router.get('/search', (req, res, next) => {
@@ -55,36 +54,24 @@ router.get('/lookup', (req, res, next) => {
 
 router.get('/signature', (req, res, next) => {
   const uuid = req.query.uuid;
-  Signature.uuid(uuid).then(signature => {
+  Signature.byUuid(uuid).then(signature => {
       res.json(signature);
   }).catch(e => {
       next(e);
   })
 });
 
-router.get('/signature/compare', (req, res, next) => {
-  const uuid1 = req.query.uuid1;
-  const uuid2 = req.query.uuid2;
-  Signature.compare(uuid1, uuid2).then(comparison => {
-      res.json(comparison);
-  }).catch(e => {
-      next(e);
-  })
-});
+router.get('/signature/:uuidCsv', (req, res, next) => {
+  const uuidCsv = req.params.uuidCsv;
+  const extraUuid = req.query.uuid;
 
-router.get('/signature2', (req, res, next) => {
-  const uuid = req.query.uuid;
-  Signature2.byUuid(uuid).then(signature => {
-      res.json(signature);
-  }).catch(e => {
-      next(e);
-  })
-});
+  const uuids = (uuidCsv !== undefined && uuidCsv !== '')? uuidCsv.split(',') : [];
+  if (extraUuid !== undefined && extraUuid !== '') {
+    uuids.push(extraUuid);
+  }
+  debug(`/signature/:uuidCsv : uuidCsv=${uuidCsv}, extraUuid=${extraUuid}, uuids=${JSON.stringify(uuids)}`);
 
-router.get('/signature2/compare', (req, res, next) => {
-  const uuid1 = req.query.uuid1;
-  const uuid2 = req.query.uuid2;
-  Signature2.byUuids([uuid1, uuid2]).then(comparison => {
+  Signature.byUuids(uuids).then(comparison => {
       res.json(comparison);
   }).catch(e => {
       next(e);
