+++ views/index.hbs
@@ -11,15 +11,8 @@
 <h2>Signature Stuff</h2>
 <ul>
   <li><a href="/articles/signature?uuid=2ebe9c54-d82e-11e7-a039-c64b1c09b482">signature</a> for <a href="https://www.ft.com/content/2ebe9c54-d82e-11e7-a039-c64b1c09b482">2ebe9c54-d82e-11e7-a039-c64b1c09b482</a></li>
-  <li><a href="/articles/signature/compare?uuid1=2ebe9c54-d82e-11e7-a039-c64b1c09b482&uuid2=d068d0b8-d529-11e7-8c9a-d9c0a5c8d5c9">compare signatures</a> for related articles, <a href="https://www.ft.com/content/2ebe9c54-d82e-11e7-a039-c64b1c09b482">2ebe9c54-d82e-11e7-a039-c64b1c09b482</a> and <a href="https://www.ft.com/content/d068d0b8-d529-11e7-8c9a-d9c0a5c8d5c9">d068d0b8-d529-11e7-8c9a-d9c0a5c8d5c9</a></li>
-  <li><a href="/articles/signature/compare?uuid1=2ebe9c54-d82e-11e7-a039-c64b1c09b482&uuid2=e535b17c-d8ea-11e7-a039-c64b1c09b482">compare signatures</a> for unrelated articles, <a href="https://www.ft.com/content/2ebe9c54-d82e-11e7-a039-c64b1c09b482">2ebe9c54-d82e-11e7-a039-c64b1c09b482</a> and <a href="https://www.ft.com/content/e535b17c-d8ea-11e7-a039-c64b1c09b482">e535b17c-d8ea-11e7-a039-c64b1c09b482</a></li>
-</ul>
-
-<h2>Signature2 Stuff</h2>
-<ul>
-  <li><a href="/articles/signature2?uuid=2ebe9c54-d82e-11e7-a039-c64b1c09b482">signature</a> for <a href="https://www.ft.com/content/2ebe9c54-d82e-11e7-a039-c64b1c09b482">2ebe9c54-d82e-11e7-a039-c64b1c09b482</a></li>
-  <li><a href="/articles/signature2/compare?uuid1=2ebe9c54-d82e-11e7-a039-c64b1c09b482&uuid2=d068d0b8-d529-11e7-8c9a-d9c0a5c8d5c9">compare signatures</a> for related articles, <a href="https://www.ft.com/content/2ebe9c54-d82e-11e7-a039-c64b1c09b482">2ebe9c54-d82e-11e7-a039-c64b1c09b482</a> and <a href="https://www.ft.com/content/d068d0b8-d529-11e7-8c9a-d9c0a5c8d5c9">d068d0b8-d529-11e7-8c9a-d9c0a5c8d5c9</a></li>
-  <li><a href="/articles/signature2/compare?uuid1=2ebe9c54-d82e-11e7-a039-c64b1c09b482&uuid2=e535b17c-d8ea-11e7-a039-c64b1c09b482">compare signatures</a> for unrelated articles, <a href="https://www.ft.com/content/2ebe9c54-d82e-11e7-a039-c64b1c09b482">2ebe9c54-d82e-11e7-a039-c64b1c09b482</a> and <a href="https://www.ft.com/content/e535b17c-d8ea-11e7-a039-c64b1c09b482">e535b17c-d8ea-11e7-a039-c64b1c09b482</a></li>
+  <li><a href="/articles/signature/2ebe9c54-d82e-11e7-a039-c64b1c09b482,d068d0b8-d529-11e7-8c9a-d9c0a5c8d5c9">compare signatures</a> for related articles, <a href="https://www.ft.com/content/2ebe9c54-d82e-11e7-a039-c64b1c09b482">2ebe9c54-d82e-11e7-a039-c64b1c09b482</a> and <a href="https://www.ft.com/content/d068d0b8-d529-11e7-8c9a-d9c0a5c8d5c9">d068d0b8-d529-11e7-8c9a-d9c0a5c8d5c9</a></li>
+  <li><a href="/articles/signature/2ebe9c54-d82e-11e7-a039-c64b1c09b482?uuid=e535b17c-d8ea-11e7-a039-c64b1c09b482">compare signatures</a> for unrelated articles, <a href="https://www.ft.com/content/2ebe9c54-d82e-11e7-a039-c64b1c09b482">2ebe9c54-d82e-11e7-a039-c64b1c09b482</a> and <a href="https://www.ft.com/content/e535b17c-d8ea-11e7-a039-c64b1c09b482">e535b17c-d8ea-11e7-a039-c64b1c09b482</a></li>
 </ul>
 
 <h2>Simple Search and Article look up</h2>
